Added DHWeaponRestrictionData class and setup basic weapon qualification check
Notes:
I wanted the system to use its own INI file. This is simply to prevent having to constantly re-qualify if you reset your config files (DarkestHour.ini and DarkestHourUser.ini). To have a custom INI file you need a way to access it, meaning you need to create an instance of it. I have DHPlayer do that, but only on the client.